rename client.ts
